Separate classes are moved to the modules package
