+++ requirements.txt
@@ -2,5 +2,5 @@ pycodestyle==2.4.0
 nose==1.3.7
 coverage==4.5.1
 termcolor==1.1.0
-requests~=2.25.1
-feedparser~=6.0.2
+requests==2.25.1
+feedparser==6.0.2
